Update email classification and SignalR integration

+++ backend/src/EmailClassification.Infrastructure/Service/BackgroundJobInitializer.cs
@@ -1,11 +1,7 @@
 ﻿using EmailClassification.Application.Interfaces.Background;
 using Hangfire;
+using Hangfire.Storage;
 using Microsoft.Extensions.Hosting;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace EmailClassification.Infrastructure.Service
 {
@@ -13,11 +9,22 @@ namespace EmailClassification.Infrastructure.Service
     {
         public Task StartAsync(CancellationToken cancellationToken)
         {
-            RecurringJob.AddOrUpdate<IBackgroundService>("sync-emails", s => s.SyncAllUsersEmails(), "*/2 * * * *");
-            RecurringJob.AddOrUpdate<IBackgroundService>("classify-emails", s => s.ClassifyAllUsersEmails(), "*/2 * * * *");
-            RecurringJob.AddOrUpdate<IBackgroundService>("delete-guest", s => s.DeleteGuestAsync(), "* * */3 * *");
-            return Task.CompletedTask;
+            var connection = JobStorage.Current.GetConnection();
+            var existingJobs = connection.GetRecurringJobs();
 
+            if (!existingJobs.Any(j => j.Id == "sync-emails"))
+            {
+                RecurringJob.AddOrUpdate<IBackgroundService>("sync-emails", s => s.SyncAllUsersEmails(), "* */1 * * *");
+            }
+            //if (!existingJobs.Any(j => j.Id == "classify-emails"))
+            //{
+            //    RecurringJob.AddOrUpdate<IBackgroundService>("classify-emails", s => s.ClassifyAllUsersEmails(), "* */1 * * *");
+            //}
+            if (!existingJobs.Any(j => j.Id == "delete-guest"))
+            {
+                RecurringJob.AddOrUpdate<IBackgroundService>("delete-guest", s => s.DeleteGuestAsync(), "* * */3 * *");
+            }
+            return Task.CompletedTask;
         }
 
         public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;
